Moved donut grid/list to own component. Passing through function + text for donut
